+++ stars-core/src/main/kotlin/tools/aqua/stars/core/metric/metrics/postEvaluation/FailedMonitorsMetric.kt
@@ -18,13 +18,15 @@
 package tools.aqua.stars.core.metric.metrics.postEvaluation
 
 import java.util.logging.Logger
-import tools.aqua.stars.core.metric.metrics.evaluation.ValidTSCInstancesPerProjectionMetric
+import tools.aqua.stars.core.metric.metrics.evaluation.ValidTSCInstancesPerTSCMetric
 import tools.aqua.stars.core.metric.providers.Loggable
 import tools.aqua.stars.core.metric.providers.PostEvaluationMetricProvider
+import tools.aqua.stars.core.metric.utils.ApplicationConstantsHolder.CONSOLE_INDENT
+import tools.aqua.stars.core.metric.utils.ApplicationConstantsHolder.CONSOLE_SEPARATOR
+import tools.aqua.stars.core.tsc.TSC
 import tools.aqua.stars.core.tsc.TSCFailedMonitorInstance
 import tools.aqua.stars.core.tsc.instance.*
 import tools.aqua.stars.core.tsc.node.TSCNode
-import tools.aqua.stars.core.tsc.projection.TSCProjection
 import tools.aqua.stars.core.types.*
 
 /**
@@ -32,15 +34,15 @@ import tools.aqua.stars.core.types.*
  * [TSCNode.monitorFunction]s that evaluate to 'false'.
  *
  * This class implements the [Loggable] interface. It logs and prints the count and names of all
- * failing [TSCNode.monitorFunction]s for each [TSCProjection]. It logs the failing
- * [TSCFailedMonitorInstance]s for each [TSCProjection].
+ * failing [TSCNode.monitorFunction]s for each [TSC]. It logs the failing
+ * [TSCFailedMonitorInstance]s for each [TSC].
  *
  * @param E [EntityType].
  * @param T [TickDataType].
  * @param S [SegmentType].
  * @param U [TickUnit].
  * @param D [TickDifference].
- * @property dependsOn The instance of a [ValidTSCInstancesPerProjectionMetric] on which this metric
+ * @property dependsOn The instance of a [ValidTSCInstancesPerTSCMetric] on which this metric
  *   depends on and needs for its calculation.
  * @property logger [Logger] instance.
  */
@@ -51,17 +53,17 @@ class FailedMonitorsMetric<
     S : SegmentType<E, T, S, U, D>,
     U : TickUnit<U, D>,
     D : TickDifference<D>>(
-    override val dependsOn: ValidTSCInstancesPerProjectionMetric<E, T, S, U, D>,
+    override val dependsOn: ValidTSCInstancesPerTSCMetric<E, T, S, U, D>,
     override val logger: Logger = Loggable.getLogger("failed-monitors")
 ) : PostEvaluationMetricProvider<E, T, S, U, D>, Loggable {
 
   /** Holds all failed monitors after calling [postEvaluate]. */
   val failedMonitors:
-      MutableMap<TSCProjection<E, T, S, U, D>, List<TSCFailedMonitorInstance<E, T, S, U, D>>> =
+      MutableMap<TSC<E, T, S, U, D>, List<TSCFailedMonitorInstance<E, T, S, U, D>>> =
       mutableMapOf()
   /**
-   * Calculates a [Map] of [TSCFailedMonitorInstance]s for all [TSCProjection]s by validating all
-   * monitors for all valid [TSCInstance]s.
+   * Calculates a [Map] of [TSCFailedMonitorInstance]s for all [TSC]s by validating all monitors for
+   * all valid [TSCInstance]s.
    */
   override fun postEvaluate() {
     failedMonitors.clear()
@@ -75,14 +77,16 @@ class FailedMonitorsMetric<
         })
   }
 
-  /** Prints the count of failed monitors for each [TSCProjection]. */
+  /** Prints the count of failed monitors for each [TSC]. */
   override fun printPostEvaluationResult() {
-    failedMonitors.forEach { (projection, failedMonitors) ->
-      logInfo("Count of failed monitors for projection '$projection': ${failedMonitors.size}")
+    println(
+        "\n$CONSOLE_SEPARATOR\n$CONSOLE_INDENT Count Of Failed Monitors Per TSC \n$CONSOLE_SEPARATOR")
+    failedMonitors.forEach { (tsc, failedMonitors) ->
+      logInfo("Count of failed monitors for tsc '${tsc.identifier}': ${failedMonitors.size}")
 
       if (failedMonitors.isEmpty()) return@forEach
 
-      logFine("Failed monitors for projection '$projection':")
+      logFine("Failed monitors for tsc '${tsc.identifier}':")
       failedMonitors.forEach { failedMonitor ->
         logFine(
             "Monitors ${failedMonitor.monitorLabel} failed in: ${failedMonitor.segmentIdentifier}")